+++ __mocks__/selenium-standalone.ts
@@ -1,7 +1,13 @@
 import { vi } from 'vitest'
-export const install = vi.fn(() => new Promise<void>((resolve) => resolve()))
-export const start = vi.fn(() => new Promise((resolve) => resolve({
-    kill : vi.fn(),
-    stdout : { pipe: vi.fn().mockReturnValue({ pipe : vi.fn() }) },
-    stderr : { pipe: vi.fn().mockReturnValue({ pipe : vi.fn() }) }
+
+const install = vi.fn(() => new Promise<void>((resolve) => resolve()))
+const start = vi.fn(() => new Promise((resolve) => resolve({
+    kill: vi.fn(),
+    stdout: { pipe : vi.fn().mockReturnValue({ pipe : vi.fn() }) },
+    stderr: { pipe : vi.fn().mockReturnValue({ pipe : vi.fn() }) }
 })))
+
+export default {
+    install,
+    start
+}