+++ packages/wdio-browser-runner/a.js
@@ -1,0 +1,3 @@
+import { resolve } from 'import-meta-resolve'
+
+console.log(await resolve('aria-query', import.meta.url))